Break out git diff from Git.ts
